Basic implementation of rotors and the Engima machine
As it stands, the output is incorrect for all things except the specific scenario of the rotors in the AAZ setting, and the rotors being one of the first three possible.

It's a good start, though. You can now press a button on the Enigma machine. Further, pressing buttons can be done either through individual characters or through inputting a string to encrypt or decrypt.